Decouple clock tracker from trace processor context to improve reusability (#2972)
Decouple the TraceProcessorContext from the ClockTracker so that the
clock tracker can be reused outside trace processor as a utility for
trace processing.
