+++ src/trace_processor/BUILD.gn
@@ -200,6 +200,7 @@ if (enable_perfetto_trace_processor_sqlite) {
       "trace_summary",
       "trace_summary:gen_cc_trace_summary_descriptor",
       "types",
+      "util:clock",
       "util:gzip",
       "util:json_parser",
       "util:protozero_to_json",
