+++ src/trace_processor/importers/common/BUILD.gn
@@ -103,6 +103,7 @@ source_set("common") {
     "../../tables:tables",
     "../../types",
     "../../util:build_id",
+    "../../util:clock",
     "../../util:profiler_util",
     "../../util:trace_type",
     "../../util/elf",
